chore:(deploy) Refactor database config for heroku
