+++ package.json
@@ -21,7 +21,9 @@
     "clean": "rimraf dist",
     "build": "babel ./server --out-dir dist",
     "prod": "cross-env NODE_ENV=production npm-run-all clean build start:prod",
-    "start": "npm-run-all db start:prod"
+    "start": "npm-run-all prod:create start:prod",
+    "prod:drop": "node dist/migrations/dropTable",
+    "prod:create": "node dist/migrations/createTable"
   },
   "repository": {
     "type": "git",
