+++ test/booking.js
@@ -2,7 +2,7 @@ import { describe, it } from 'mocha';
 import chai,{expect} from 'chai';
 import chaiHttp from 'chai-http';
 
-import app from '../app';
+import app from '../server/app';
 
 chai.should();
